refactor: category Page, index Page 추가
